+++ src/lib.rs
@@ -51,9 +51,6 @@
 //! - Module `event`
 //!   - Mouse events - [`EnableMouseCapture`](event/struct.EnableMouseCapture.html),
 //!     [`DisableMouseCapture`](event/struct.DisableMouseCapture.html)
-//! - Module `screen`
-//!   - Alternate screen - [`EnterAlternateScreen`](screen/struct.EnterAlternateScreen.html),
-//!     [`LeaveAlternateScreen`](screen/struct.LeaveAlternateScreen.html)
 //! - Module `style`
 //!   - Colors - [`SetForegroundColor`](style/struct.SetForegroundColor.html),
 //!     [`SetBackgroundColor`](style/struct.SetBackgroundColor.html),
@@ -65,6 +62,8 @@
 //!     [`ScrollDown`](terminal/struct.ScrollDown.html)
 //!   - Miscellaneous - [`Clear`](terminal/struct.Clear.html),
 //!     [`SetSize`](terminal/struct.SetSize.html)
+//!   - Alternate screen - [`EnterAlternateScreen`](screen/struct.EnterAlternateScreen.html),
+//!     [`LeaveAlternateScreen`](screen/struct.LeaveAlternateScreen.html)
 //!
 //! ### Command Execution
 //!
@@ -228,16 +227,12 @@
 //! [stderr]: https://doc.rust-lang.org/std/io/fn.stderr.html
 //! [flush]: https://doc.rust-lang.org/std/io/trait.Write.html#tymethod.flush
 
-#[cfg(windows)]
-pub use utils::functions::supports_ansi;
-pub use utils::{Command, ErrorKind, ExecutableCommand, Output, QueueableCommand, Result};
+pub use utils::{Command, ErrorKind, ExecutableCommand, QueueableCommand, Result};
 
 /// A module to work with the terminal cursor
 pub mod cursor;
 /// A module to read events.
 pub mod event;
-/// A module to work with the terminal screen and modes.
-pub mod screen;
 /// A module to apply attributes and colors on your text.
 pub mod style;
 /// A module to work with the terminal.