+++ src/utils/command.rs
@@ -1,43 +1,53 @@
 use std::{fmt::Display, io::Write};
 
-use crate::{execute, queue, write_cout};
+use crate::{execute, queue};
 
 use super::error::Result;
 
-/// A command is an action that can be performed on the terminal.
+/// An interface for a command that performs an action on the terminal.
 ///
-/// crossterm already delivers a number of commands.
-/// There is no need to implement them yourself.
-/// Also, you don't have to execute the commands yourself by calling a function.
+/// Crossterm provides a set of commands,
+/// and there is no immediate reason to implement a command yourself.
+/// In order to understand how to use and execute commands,
+/// it is recommended that you take a look at [Command Api](../#command-api) chapter.
 pub trait Command {
     type AnsiType: Display;
 
-    /// Returns the ANSI code representation of this command.
-    /// You can manipulate the terminal behaviour by writing an ANSI escape code to the terminal.
-    /// You are able to use ANSI escape codes only for windows 10 and UNIX systems.
+    /// Returns an ANSI code representation of this command.
+    /// An ANSI code can manipulate the terminal by writing it to the terminal buffer.
+    /// However, only Windows 10 and UNIX systems support this.
     ///
-    /// **This method is mainly used internally by crossterm!**
+    /// This method does not need to be accessed manually, as it is used by the crossterm's [Command Api](../#command-api)
     fn ansi_code(&self) -> Self::AnsiType;
 
     /// Execute this command.
     ///
-    /// On operating systems that do not support ANSI escape codes ( < Windows 10) we need to call WinApi to execute this command.
+    /// Windows versions lower than windows 10 do not support ANSI escape codes,
+    /// therefore a direct WinAPI call is made.
     ///
-    /// **This method is mainly used internally by crossterm!**
+    /// This method does not need to be accessed manually, as it is used by the crossterm's [Command Api](../#command-api)
     #[cfg(windows)]
     fn execute_winapi(&self) -> Result<()>;
+
+    /// Returns whether the ansi code representation of this command is supported by windows.
+    ///
+    /// A list of supported ANSI escape codes
+    /// can be found [here](https://docs.microsoft.com/en-us/windows/console/console-virtual-terminal-sequences).
+    #[cfg(windows)]
+    fn is_ansi_code_supported(&self) -> bool {
+        super::functions::supports_ansi()
+    }
 }
 
-/// A trait that defines behaviour for a command which can be executed at a later time.
-/// This can be used in order to get more performance.
+/// An interface for commands that can be queued for further execution.
 pub trait QueueableCommand<T: Display>: Sized {
-    /// Queues the given command for later execution.
+    /// Queues the given command for further execution.
     fn queue(&mut self, command: impl Command<AnsiType = T>) -> Result<&mut Self>;
 }
 
-/// A trait that defines behaviour for a command which will be executed immediately.
+/// An interface for commands that are directly executed.
 pub trait ExecutableCommand<T: Display>: Sized {
-    /// Execute the given command directly.
+    /// Executes the given command directly.
     fn execute(&mut self, command: impl Command<AnsiType = T>) -> Result<&mut Self>;
 }
 
@@ -46,24 +56,57 @@ where
     A: Display,
     T: Write,
 {
-    /// Queue the given command for later execution.
+    /// Queues the given command for further execution.
     ///
     /// Queued commands will be executed in the following cases:
-    /// - When you manually call `flush` on the given writer.
-    /// - When the buffer is to full, then the terminal will flush for you.
-    /// - Incase of `stdout` each line, because `stdout` is line buffered.
     ///
-    /// # Parameters
+    /// * When `flush` is called manually on the given type implementing `io::Write`.
+    /// * The terminal will `flush` automatically if the buffer is full.
+    /// * Each line is flushed in case of `stdout`, because it is line buffered.
+    ///
+    /// # Arguments
+    ///
     /// - [Command](./trait.Command.html)
     ///
     ///     The command that you want to queue for later execution.
     ///
-    /// # Remarks
-    /// - In the case of UNIX and windows 10, ANSI codes are written to the given 'writer'.
-    /// - In case of Windows versions lower than 10, a direct WinApi call will be made.
-    /// This is happening because windows versions lower then 10 do not support ANSI codes, and thus they can't be written to the given buffer.
-    /// Because of that there is no difference between `execute` and `queue` for those windows versions.
-    /// - Queuing might sound that there is some scheduling going on, however, this means that we write to the stdout without flushing which will cause commands to be stored in the buffer without them being written to the terminal.
+    /// # Examples
+    ///
+    /// ```rust
+    /// use std::io::{Write, stdout};
+    /// use crossterm::{Result, QueueableCommand, style::Print};
+    ///
+    ///  fn main() -> Result<()> {
+    ///     let mut stdout = stdout();
+    ///
+    ///     // `Print` will executed executed when `flush` is called.
+    ///     stdout
+    ///         .queue(Print("foo 1\n".to_string()))?
+    ///         .queue(Print("foo 2".to_string()))?;
+    ///
+    ///     // some other code (no execution happening here) ...
+    ///
+    ///     // when calling `flush` on `stdout`, all commands will be written to the stdout and therefore executed.
+    ///     stdout.flush()?;
+    ///
+    ///     Ok(())
+    ///
+    ///     // ==== Output ====
+    ///     // foo 1
+    ///     // foo 2
+    /// }
+    /// ```
+    ///
+    /// Have a look over at the [Command API](./#command-api) for more details.
+    ///
+    /// # Notes
+    ///
+    /// * In the case of UNIX and Windows 10, ANSI codes are written to the given 'writer'.
+    /// * In case of Windows versions lower than 10, a direct WinApi call will be made.
+    ///     The reason for this is that Windows versions lower than 10 do not support ANSI codes,
+    ///     and can therefore not be written to the given `writer`.
+    ///     Therefore, there is no difference between [execute](./trait.ExecutableCommand.html)
+    ///     and [queue](./trait.QueueableCommand.html) for those old Windows versions.
     fn queue(&mut self, command: impl Command<AnsiType = A>) -> Result<&mut Self> {
         queue!(self, command)?;
         Ok(self)
@@ -75,47 +118,48 @@ where
     A: Display,
     T: Write,
 {
-    /// Execute the given command directly.
-    /// This function will `write` the ANSI escape code to this type and call `flush`.
+    /// Executes the given command directly.
     ///
-    /// In case you have many executions after on and another you can use `queue(command)` to get some better performance.
-    /// The `queue` function will not call `flush`.
+    /// The given command its ANSI escape code will be written and flushed onto `Self`.
     ///
-    /// # Remarks
-    /// - In the case of UNIX and windows 10, ANSI codes are written to the given 'writer'.
-    /// - In case of Windows versions lower than 10, a direct WinApi call will be made.
-    /// This is happening because Windows versions lower then 10 do not support ANSI codes, and thus they can't be written to the given buffer.
-    /// Because of that there is no difference between `execute` and `queue` for those windows versions.
+    /// # Arguments
+    ///
+    /// - [Command](./trait.Command.html)
+    ///
+    ///     The command that you want to execute directly.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use std::io::{Write, stdout};
+    /// use crossterm::{Result, ExecutableCommand, style::Print};
+    ///
+    ///  fn main() -> Result<()> {
+    ///      // will be executed directly
+    ///       stdout()
+    ///         .execute(Print("sum:\n".to_string()))?
+    ///         .execute(Print(format!("1 + 1= {} ", 1 + 1)))?;
+    ///
+    ///       Ok(())
+    ///
+    ///      // ==== Output ====
+    ///      // sum:
+    ///      // 1 + 1 = 2
+    ///  }
+    /// ```
+    ///
+    /// Have a look over at the [Command API](./#command-api) for more details.
+    ///
+    /// # Notes
+    ///
+    /// * In the case of UNIX and Windows 10, ANSI codes are written to the given 'writer'.
+    /// * In case of Windows versions lower than 10, a direct WinApi call will be made.
+    ///     The reason for this is that Windows versions lower than 10 do not support ANSI codes,
+    ///     and can therefore not be written to the given `writer`.
+    ///     Therefore, there is no difference between [execute](./trait.ExecutableCommand.html)
+    ///     and [queue](./trait.QueueableCommand.html) for those old Windows versions.
     fn execute(&mut self, command: impl Command<AnsiType = A>) -> Result<&mut Self> {
         execute!(self, command)?;
         Ok(self)
-    }
-}
-
-/// When executed, this command will output the given displayable to the buffer.
-///
-/// See `crossterm/examples/command.rs` for more information on how to execute commands.
-pub struct Output<T: Display + Clone>(pub T);
-
-impl<T: Display + Clone> Command for Output<T> {
-    type AnsiType = T;
-
-    fn ansi_code(&self) -> Self::AnsiType {
-        self.0.clone()
-    }
-
-    #[cfg(windows)]
-    fn execute_winapi(&self) -> Result<()> {
-        print!("{}", self.0);
-        Ok(())
-    }
-}
-
-impl<T: Display + Clone> Display for Output<T> {
-    fn fmt(
-        &self,
-        f: &mut ::std::fmt::Formatter<'_>,
-    ) -> ::std::result::Result<(), ::std::fmt::Error> {
-        write!(f, "{}", self.ansi_code())
     }
 }